api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["ngoclinh <ngoclinh8123@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.6"
django = "^4.1.5"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
